chore: organize type exports
- add barrel file to export all palette-types declarations under one
  'palette-types' alias for cleaner imports
- update directory names to be consistent across the project
- export local types to types directory
- update components to use new alias import
- note: the relative path will still work
- TG-216
