+++ frontend/src/features/rubricBuilder/CriteriaInput.tsx
@@ -10,7 +10,7 @@ import { useSortable } from "@dnd-kit/sortable"; // Import useSortable
 import { CSS } from "@dnd-kit/utilities"; // Import CSS utilities
 
 import RatingInput from "./RatingInput.tsx";
-import { Criteria, Rating } from "../../../../palette-types/src";
+import { Criteria, Rating } from "palette-types";
 import { calcMaxPoints } from "../../utils/calculateMaxPoints.ts";
 import { createRating } from "../../utils/rubricFactory.ts";
