+++ frontend/src/types/csvRow.ts
@@ -1,0 +1,2 @@
+// define our csv rows for the data field in papa parse
+export type CSVRow = [string, ...(number | string)[]];
