--- frontend/src/types/placeholder.txt
@@ -1,1 +1,0 @@
-This directory will contain frontend specific types such as component props.
