+++ palette-types/src/canvasProtocol/DeleteRubricRequest.ts
