+++ palette-types/src/canvasProtocol/GetRubricRequest.ts
