+++ palette-types/src/canvasTypes/CanvasAssessment.ts
