+++ palette-types/src/canvasTypes/CanvasAssociation.ts
