+++ palette-types/src/canvasTypes/CanvasCriterion.ts
