+++ palette-types/src/canvasTypes/CanvasRating.ts
