+++ palette-types/src/canvasTypes/CanvasRubric.ts
