+++ palette-types/src/protocol/PaletteAPIErrorData.ts
