+++ palette-types/src/protocol/PaletteAPIRequest.ts
