+++ palette-types/src/protocol/PaletteAPIResponse.ts
