+++ palette-types/src/types/Criteria.ts
