+++ palette-types/src/types/Rating.ts
