+++ palette-types/src/types/Rubric.ts
